Organise notebooks and files
