+++ .gitignore
@@ -1,3 +1,4 @@
+# Folders with data and trainig data
 raw_data/
 data/
 training_outputs/
